immichUploader/requirements.txt
python-dotenv
tqdm